[FIX] types in YCbCr

+++ include/ycbcr2rgb.h
@@ -9,7 +9,7 @@
 #include <utils.h>
 
 // Fonction pour saturer les valeurs entre 0 et 255
-uint8_t saturer(int valeur);
+uint8_t saturer(int16_t valeur);
 
 // Fonction pour convertir un pixel YCbCr en pixel RGB
 // Si on a une seule composante, on ne met à jour que la composante 0
